requirements.txt
elasticsearch==7.17.6
elasticsearch_dsl==7.4.0
frontend==0.0.3
PyMuPDF==1.19.0
Flask==2.1.0
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
python-dotenv
elastic-apm[flask]
blinker
gunicorn==20.1.0
Authlib==1.2.0